requirements.txt
pyyaml
requests
twitchAPI
